Add dry-run flag and fix some issues
